Redo modal display control
Modal displays for clicks on graph nodes now handled in the <Graph> component — the first step towards separating the graph modal displays from <FileGalleryRenderer>. ENCD-3510
